tests/test-requirements.txt
pytest>=7.4.0
pytest-flask>=1.2.0

